structure changed: /modules renamed to /components, index and style file added to each component
