+++ react-cat/src/components/About/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 function About() {
-    return <h1>About page</h1>;
+    return <h1>Index page</h1>;
 }
 
 export default About;
