+++ react-cat/src/components/Main/style.css
