+++ react-cat/src/components/MainContent/style.css
