+++ react-cat/src/components/ReposList/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./style.css";
+
+class RepoList extends React.Component {
+    render() {
+        return <h1>Repos List</h1>;
+    }
+}
+
+export default RepoList;
