+++ react-cat/src/components/UserInfo/style.css
