--- react-cat/src/modules/repos-list/ReposList.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./ReposList.css";
-
-class ReposList extends React.Component {
-    render() {
-        return <h1>Repos List</h1>;
-    }
-}
-
-export default ReposList;
